aider: Add --fast command line argument to skip time.sleep.
# Aider chat conversation:

USER: Add another command line argument, --fast that, when set, does not time.sleep

ASSISTANT: To add the `--fast` command line argument and modify the behavior of the `time.sleep` call based on this argument, we need to update the `main.py` file. The `--fast` argument will be a flag that, when set, will skip the `time.sleep` call to speed up the process.

Here's how to add the `--fast` argument and modify the `time.sleep` call:

```python
main.py
<<<<<<< SEARCH
    parser.add_argument("--skip-chapters", type=str, help="Comma separated list of chapters to exclude")
=======
    parser.add_argument("--skip-chapters", type=str, help="Comma separated list of chapters to exclude")
    parser.add_argument("--fast", action="store_true", help="Run without artificial delay between chunks")
>>>>>>> REPLACE
```

And for the `time.sleep` modification:

```python
main.py
<<<<<<< SEARCH
        logger.info("Chunk written, waiting 30s")
        time.sleep(30)
=======
        if not args.fast:
            logger.info("Chunk written, waiting 30s")
            time.sleep(30)
>>>>>>> REPLACE
```

With these changes, the `--fast` argument will be recognized by the argument parser, and the `time.sleep` call will be conditional based on whether `--fast` is set or not. If `--fast` is set, the script will not wait between processing chunks.